+++ test/realtimecontroller/DefaultRealtimeController.test.ts
@@ -548,6 +548,16 @@ describe('DefaultRealtimeController', () => {
       expect(state.volumeIndicatorCallbacks).to.equal(expectedResult);
     });
 
+    it('will not error if unsubscribed from a non-exist attendeeId', () => {
+      const callback = (
+        _attendeeId: string,
+        _volume: number | null,
+        _muted: boolean | null,
+        _signalStrength: number | null
+      ): void => {};
+      rt.realtimeUnsubscribeFromVolumeIndicator('bar-attendee', callback);
+    });
+
     it('will tolerate an exception thrown in a volume indicator callback', () => {
       const sentAttendeeId = 'foo-attendee';
       const sentVolume = 0.5;
@@ -1192,29 +1202,6 @@ describe('DefaultRealtimeController', () => {
         "Cannot read property 'indexOf' of undefined",
         "Cannot read properties of undefined (reading 'indexOf')"
       );
-      expect(fatal.calledWith(matchUn)).to.be.true;
-    });
-
-    it('handles broken volume callbacks', () => {
-      const fatal = sinon.stub();
-      rt.realtimeSubscribeToFatalError(fatal);
-
-      // Break it.
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const state: RealtimeState = ((rt as unknown) as any).state as RealtimeState;
-      state.volumeIndicatorCallbacks = undefined;
-
-      rt.realtimeSubscribeToVolumeIndicator('a', (_a, _v, _m, _s) => {});
-      expect(fatal.calledOnce).to.be.true;
-      const match = matchError(
-        "Cannot read property 'hasOwnProperty' of undefined",
-        "Cannot read properties of undefined (reading 'hasOwnProperty')"
-      );
-      expect(fatal.calledWith(match)).to.be.true;
-      fatal.reset();
-
-      rt.realtimeUnsubscribeFromVolumeIndicator('a');
-      const matchUn = matchError('Cannot convert undefined or null to object');
       expect(fatal.calledWith(matchUn)).to.be.true;
     });